+++ packages/editor/src/components/post-actions/set-as-homepage.js
@@ -9,6 +9,8 @@ import {
 	__experimentalHStack as HStack,
 	__experimentalVStack as VStack,
 } from '@wordpress/components';
+// eslint-disable-next-line @wordpress/use-recommended-components
+import { Dialog } from '@wordpress/ui';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
 import { store as noticesStore } from '@wordpress/notices';
@@ -94,17 +96,9 @@ const SetAsHomepageModal = ( { items, closeModal } ) => {
 			<VStack spacing="5">
 				<WCText>{ modalText }</WCText>
 				<HStack justify="right">
-					<Button
-						__next40pxDefaultSize
-						variant="tertiary"
-						onClick={ () => {
-							closeModal?.();
-						} }
-						disabled={ isSaving }
-						accessibleWhenDisabled
-					>
+					<Dialog.Action variant="outline" disabled={ isSaving }>
 						{ __( 'Cancel' ) }
-					</Button>
+					</Dialog.Action>
 					<Button
 						__next40pxDefaultSize
 						variant="primary"